product change small fix

+++ resources/js/services/globals.js
@@ -8,8 +8,11 @@ if (location.pathname.startsWith('/checkout')) {
   const product_id = cookies.getCookie('product_id');
 
   if (product_id !== checkoutData.product.id) {
-    localStorage.clear();
-    cookies.clearCookies({ except: ['XSRF-TOKEN'] });
+    if (product_id) {
+      localStorage.clear();
+      cookies.clearCookies({ except: ['XSRF-TOKEN'] });
+    }
+
     cookies.setCookie('product_id', checkoutData.product.id);
   }
 }
